Made a rectangle
This is no ordinary rectangle. This rectangle represents one of the most important things that will exist in the game. It represents our third boss, who will probably be a wizard. It currently has a width of 2 units and a height of 3 units. While it is currently white, it will later be edited to be another color as a temporary placeholder, most likely a light blue. It does not have any scripts yet, but it will. It will. The most notable thing about this rectangle is its potential. Any rectangle could become a wizard, but only this rectangle could become the best wizard boss ever. It's really quite impressive, really. It has more potential than you will ever have. It has more potential than the president of the united states. It really is just super awesome at having potential. While on the topic of Presidents, there is one more thing it's really good at. This rectangle is super good at among us. It gets selected as the imposter 95% of the time (against all odds), which upsets all of its lame rectangle friends, and this rectangle has never lost a game. Maybe this is its  true potential being realized. It is unclear really. It is not for our minds to comprehend. Anyways, yeah. I added a rectangle. You're welcome.